requirements.txt
flask
flask-cors
openai
gunicorn
python-dotenv
pydub
ffmpeg-python
